- Decompose SearchPage logic into talon - Move query param utility to Peregrine - Replace direct usage of location and history with new hooks
